Add support for I2C_SLAVE_FORCE
Tools like i2cget or i2cset have a "-f" (force) option to allow
reading from or writing to a device even if it is busy. This commit
makes this functionality available through an extra parameter to all the
read*/write* methods. This new feature is optional and disabled by
default.